# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/interests": {
      "post": {
        "operationId": "InterestController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "get": {
        "operationId": "InterestController_getAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
...
    "/api/v1/interests/dynamic-fields": {
      "get": {
        "operationId": "InterestController_getDynamicFieldsConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "post": {
        "operationId": "InterestController_addDynamicField",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDynamicFieldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "patch": {
        "operationId": "InterestController_replaceAllDynamicFields",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDynamicFieldsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
    "/api/v1/interests/dynamic-fields/{fieldId}": {
      "delete": {
        "operationId": "InterestController_removeDynamicField",
        "parameters": [
          {
            "name": "fieldId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "patch": {
        "operationId": "InterestController_updateDynamicField",
        "parameters": [
          {
            "name": "fieldId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
...
    "/api/v1/interests/request": {
      "get": {
        "operationId": "InterestController_getInterestsByStatus",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
    "/api/v1/interests/request/{userId}": {
      "patch": {
        "operationId": "InterestController_updateUserStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
    "/api/v1/interests/by-id/{id}": {
      "get": {
        "operationId": "InterestController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "delete": {
        "operationId": "InterestController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
    "/api/v1/interests/by-email/{email}": {
      "get": {
        "operationId": "InterestController_getByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      },
      "patch": {
        "operationId": "InterestController_update",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInterestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Interests"
        ]
      }
    },
...
    "/api/v1/microgrant/form": {
      "post": {
        "operationId": "MicroGrantController_createOrUpdateForm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateFormDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      },
      "get": {
        "operationId": "MicroGrantController_getForm",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
    "/api/v1/microgrant/apply": {
      "post": {
        "operationId": "MicroGrantController_applyForGrant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyMicroGrantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
    "/api/v1/microgrant/applications": {
      "get": {
        "operationId": "MicroGrantController_getApplications",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
    "/api/v1/microgrant/application/{userId}": {
      "get": {
        "operationId": "MicroGrantController_getUserApplication",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
    "/api/v1/microgrant/application/{id}/status": {
      "patch": {
        "operationId": "MicroGrantController_updateApplicationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
    "/api/v1/microgrant/application/check/{userId}": {
      "get": {
        "operationId": "MicroGrantController_checkApplication",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Micro Grant"
        ]
      }
    },
...
  },
  "info": {
    "title": "CCC Backend API",
    "description": "Christ Church Commission (CCC) Pastor Mentor platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "ChurchDetailsDto": {
        "type": "object",
        "properties": {
          "churchName": {
            "type": "string"
          },
          "churchPhone": {
            "type": "string"
          },
          "churchWebsite": {
            "type": "string"
          },
          "churchAddress": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "CreateInterestDto": {
        "type": "object",
        "properties": {
          "profileInfo": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdBy": {
            "enum": [
              "self",
              "admin"
            ],
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "churchDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChurchDetailsDto"
            }
          },
          "title": {
            "type": "string",
            "enum": [
              "Pastor",
              "Lay Leader",
              "Seminarian",
              "Mentor",
              "Field Mentor",
              "Director"
            ]
          },
          "conference": {
            "type": "string"
          },
          "yearsInMinistry": {
            "type": "string"
          },
          "currentCommunityProjects": {
            "type": "string"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comments": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "dynamicFieldValues": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "title"
        ]
      },
      "AddDynamicFieldDto": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "placeholder": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "order": {
            "type": "number"
          },
          "section": {
            "type": "string"
          }
        },
        "required": [
          "fieldId",
          "label",
          "type"
        ]
      },
      "DynamicFieldDto": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "placeholder": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "order": {
            "type": "number"
          },
          "section": {
            "type": "string"
          }
        },
        "required": [
          "fieldId",
          "label",
          "type"
        ]
      },
      "UpdateDynamicFieldsDto": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DynamicFieldDto"
            }
          }
        },
        "required": [
          "fields"
        ]
      },
      "UpdateInterestDto": {
        "type": "object",
        "properties": {
          "profileInfo": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdBy": {
            "enum": [
              "self",
              "admin"
            ],
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "churchDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChurchDetailsDto"
            }
          },
          "title": {
            "type": "string",
            "enum": [
              "Pastor",
              "Lay Leader",
              "Seminarian",
              "Mentor",
              "Field Mentor",
              "Director"
            ]
          },
          "conference": {
            "type": "string"
          },
          "yearsInMinistry": {
            "type": "string"
          },
          "currentCommunityProjects": {
            "type": "string"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comments": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "dynamicFieldValues": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "ChoiceDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
      "LayerDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChoiceDto"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "RecommendationLevelDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "level",
          "items"
        ]
      },
      "SectionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "layers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LayerDto"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationLevelDto"
            }
          }
        },
        "required": [
          "title",
          "description",
          "layers",
          "recommendations"
        ]
      },
...
      "CreateOrUpdateFormDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionDto"
            }
          },
          "reportingProcedure": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "sections"
        ]
      },
      "ApplyMicroGrantDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "answers": {
            "type": "object"
          },
          "supportingDoc": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "answers"
        ]
      },
      "UpdateApplicationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}